refactor token feature
