+++ hooks/token/useBurnToken.ts
@@ -1,9 +1,9 @@
 import { burn } from "@solana/spl-token";
 import { LAMPORTS_PER_SOL } from "@solana/web3.js";
 import React, { useState } from "react";
-import { useConnectionState } from "../context/ConnectionProvider";
-import { useKeypairState } from "../context/KeypairProvider";
-import { useTokenState } from "../context/TokenProvider";
+import { useConnectionState } from "../../context/ConnectionProvider";
+import { useKeypairState } from "../../context/KeypairProvider";
+import { useTokenState } from "../../context/TokenProvider";
 
 interface BurnTokenData {
   state: "error" | "start" | "finish";
